Refactor KOB service layer: split 2 monoliths into 9 focused modules
kob_scheduler.py (2030 lines) → kob_time, kob_algorithms, kob_preview, kob_suggest
kob_service.py (1700 lines) → kob_queries, kob_scoring, kob_advancement, kob_responses

Both original files kept as thin re-export facades for zero import churn
in routes, tests, and auth. All imports at module top level, no circular
dependencies.
